% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{annotateVcf,CellbaseQuery-method}
\alias{annotateVcf,CellbaseQuery-method}
\title{This method is a convience method to annoatate a vcf files}
\usage{
\S4method{annotateVcf}{CellbaseQuery}(object, file, ...)
}
\arguments{
\item{object}{An object of class CellbaseQuery}

\item{file}{Path to a bgzipped and tabix indexed vcf file}

\item{...}{any extra arguments}
}
\value{
an annotated dataframe
}
\description{
This method is a convience method to annoatate a vcf files
}
\details{
This methods is ideal for annotating small to medium sized vcf files
}

