% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Annovcf}
\alias{Annovcf}
\title{A convience fubction to directly annotate variants from a vcf file}
\usage{
Annovcf(object, file, batch_size, num_threads)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{file}{Path to a bgzipped and tabix indexed vcf file}

\item{batch_size}{intger if multiple queries are raised by a single method 
call, e.g. getting annotation info for several genes,
queries will be sent to the server in batches. This slot indicates the size
 of each batch, e.g. 200}

\item{num_threads}{integer number of asynchronus batches to be sent to the 
server}

\item{...}{any extra arguments}
}
\value{
a dataframe
}
\description{
This is a function to annotate variants from a vcf file
}

