% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbChromosomeInfoClient-methods.R
\docType{methods}
\name{cbChromosomeInfoClient,CellBaseR-method}
\alias{cbChromosomeInfoClient}
\alias{cbChromosomeInfoClient,CellBaseR-method}
\title{A method to query sequence data from Cellbase web services}
\usage{
\S4method{cbChromosomeInfoClient}{CellBaseR}(object, ids, resource,
  filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried}

\item{resource}{a character vector to specify the ids to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query sequence data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbChromosomeInfoClient(object=cb, ids="22", resource="info")
}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

