% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbClinical,CellbaseQuery-method}
\alias{cbClinical}
\alias{cbClinical,CellbaseQuery-method}
\title{A method to query Clinical data from Cellbase web services}
\usage{
\S4method{cbClinical}{CellbaseQuery}(object, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{filters}{a object of class CellbaseParam specifying the paramters 
limiting the CellbaseQuery}

\item{...}{any extra arguments}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query Clinical data from Cellbase web services
}
\examples{
library(cellbaseR)
cb <- CellbaseQuery()
cbparam <- CellbaseParam(gene=c("TP73","TET1"))
res <- cbClinical(object=cb,filters=cbParam)

}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

