% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{cbClinical}
\alias{cbClinical}
\title{A method to query Clinical data from Cellbase web services}
\usage{
cbClinical(object, filters, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{filters}{a object of class CellbaseParam specifying the paramters limiting the CellbaseQuery}

\item{...}{any extra arguments}
}
\value{
an object of class CellbaseResult which holds a dataframe
 doMC R.utils
}
\description{
A method to query Clinical data from Cellbase web services
}

