% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbClinicalClient-methods.R
\docType{methods}
\name{cbClinicalClient,CellBaseR-method}
\alias{cbClinicalClient,CellBaseR-method}
\alias{cbClinicalClientClient}
\title{A method to query Clinical data from Cellbase web services}
\usage{
\S4method{cbClinicalClient}{CellBaseR}(object, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{filters}{a object of class CellBaseParam specifying the paramters
limiting the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query Clinical data from Cellbase web services
}
\examples{
library(cellbaseR)
cb <- CellBaseR()
cbParam <- CellBaseParam(gene=c("TP73","TET1"))
res <- cbClinicalClient(object=cb,filters=cbParam)

}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

