% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cbClinicalClient}
\alias{cbClinicalClient}
\title{A method to query Clinical data from Cellbase web services. Please, for details on possible values for the parameters 
and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful API 
documentation  http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
\usage{
cbClinicalClient(object, filters, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{filters}{a object of class CellBaseParam specifying the parameters limiting the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query Clinical data from Cellbase web services. Please, for details on possible values for the parameters 
and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful API 
documentation  http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   library(cellbaseR)
   cb <- CellBaseR()
   cbParam <- CellBaseParam(gene=c("TP73","TET1"))
   res <- cbClinicalClient(object=cb,filters=cbParam)
}

