% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbData-methods.R
\docType{methods}
\name{cbData,CellBaseResponse-method}
\alias{cbData}
\alias{cbData,CellBaseResponse-method}
\title{An accessor method to get CellBaseResponse cbData slot}
\usage{
\S4method{cbData}{CellBaseResponse}(object)
}
\arguments{
\item{object}{an object of class CellBaseResponse}
}
\value{
a dataframe
}
\description{
An accessor method to get CellBaseResponse cbData slot
}
\examples{
 cb <- CellBaseR()
 res <- cbGeneClient(object=cb, ids="TET1", resource="info")
 res <- cbData(res)
}

