% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cbData}
\alias{cbData}
\title{An accessor method to get CellBaseResponse cbData slot}
\usage{
cbData(object, ...)
}
\arguments{
\item{object}{an object of class CellBaseResponse}
}
\value{
a dataframe
}
\description{
An accessor method to get CellBaseResponse cbData slot
}

