% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbGenomeSequence,CellbaseQuery-method}
\alias{cbGenomeSequence}
\alias{cbGenomeSequence,CellbaseQuery-method}
\title{A method to query sequence data from Cellbase web services}
\usage{
\S4method{cbGenomeSequence}{CellbaseQuery}(object, ids, resource,
  filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{a object of class CellbaseParam specifying additional filters
for the CellbaseQuery}

\item{...}{any extra arguments}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query sequence data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellbaseQuery()
   res <- cbGenomeSequence(object=cb, ids="22", resource="info")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

