% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{cbHelp}
\alias{cbHelp}
\title{A function to get help about cellbase queries}
\usage{
cbHelp(object, category, subcategory, resource = NULL)
}
\arguments{
\item{object}{a cellBase class object}

\item{category}{a character the category to be queried}

\item{subcategory}{a character the subcategory to be queried}

\item{resource}{A charcter when specified will get all the parametrs for
that specific resource}
}
\description{
This is a convience function to get help on cellbase methods
}
\examples{
cbHelp(cb, category="feature", subcategory="gene")
}

