% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbRegion,CellbaseQuery-method}
\alias{cbRegion}
\alias{cbRegion,CellbaseQuery-method}
\title{A method to query region data from Cellbase web services}
\usage{
\S4method{cbRegion}{CellbaseQuery}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{a object of class CellbaseParam specifying additional filterss for the CellbaseQuery}

\item{...}{any extra arguments}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query region data from Cellbase web services
}
\examples{
library(cellbaseR)
 cb <- CellbaseQuery()
 res <- cbRegion(object=cb, ids="17:1000000-1200000", resource="gene")
}
\seealso{
for more information about the cellbase webservices see 
\url{https://github.com/opencb/cellbase/wiki}
}

