% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRegionClient-methods.R
\docType{methods}
\name{cbRegionClient,CellBaseR-method}
\alias{cbRegionClient}
\alias{cbRegionClient,CellBaseR-method}
\title{A method to query region data from Cellbase web services}
\usage{
\S4method{cbRegionClient}{CellBaseR}(object, ids, resource, filters = NULL,
  ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried, for example,
"1:1000000-1200000' should always be in the form 'chr:start-end'}

\item{resource}{a character vector to specify the ids to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query region data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbRegionClient(object=cb, ids="17:1000000-1200000", resource="gene")
}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

