% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbSnpClient,CellBaseR-method}
\alias{cbSnpClient}
\alias{cbSnpClient,CellBaseR-method}
\title{A method to query snp data from Cellbase web services}
\usage{
\S4method{cbSnpClient}{CellBaseR}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a charcter vector of the ids to be queried, must be a valid rsid,
for example 'rs6025'}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query snp data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbSnpClient(object=cb, ids="rs6025", resource="info")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

