% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbSpecies,CellbaseQuery-method}
\alias{cbSpecies}
\alias{cbSpecies,CellbaseQuery-method}
\title{A method for getting the avaiable species from the cellbase web services}
\usage{
\S4method{cbSpecies}{CellbaseQuery}(object, ...)
}
\arguments{
\item{object}{An object of class CellbaseQuery}
}
\description{
A method for getting the avaiable species from the cellbase web services
}
\details{
A method for getting the avaiable species from the cellbase 
web services
}
\examples{
library(cellbaseR)
   cb <- CellbaseQuery()
   res <- cbSpecies(object=cb)
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

