% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbSpeciesClient-methods.R
\docType{methods}
\name{cbSpeciesClient,CellBaseR-method}
\alias{cbSpeciesClient}
\alias{cbSpeciesClient,CellBaseR-method}
\title{A method for getting the avaiable species from the cellbase web services}
\usage{
\S4method{cbSpeciesClient}{CellBaseR}(object, ...)
}
\arguments{
\item{object}{An object of class CellBaseR}
}
\description{
A method for getting the avaiable species from the cellbase web services
}
\details{
A method for getting the avaiable species from the cellbase
web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbSpeciesClient(object=cb)
}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

