% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbTfbsClient,CellBaseR-method}
\alias{cbTfbsClient}
\alias{cbTfbsClient,CellBaseR-method}
\title{A method to query transcription factors binding sites data from Cellbase 
web services}
\usage{
\S4method{cbTfbsClient}{CellBaseR}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a charcter vector of the ids to be queried, must be a valid 
transcription factor name, for example, Pdr1, and Oaf1}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query transcription factors binding sites data from Cellbase 
web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbTfbsClient(object=cb, ids="PAX1", resource="gene")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

