% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbTranscript,CellbaseQuery-method}
\alias{cbTranscript}
\alias{cbTranscript,CellbaseQuery-method}
\title{A method to query transcript data from Cellbase web services}
\usage{
\S4method{cbTranscript}{CellbaseQuery}(object, ids, resource, filters = NULL,
  ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the resource to be queried can
be any of "info", "function_prediction", "gene", "sequence", "variation"
, or "protein"}

\item{filters}{a object of class CellbaseParam specifying additional filters
for the CellbaseQuery}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query transcript data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellbaseQuery()
   res <- cbTranscript(object=cb, ids="ENST00000373644", resource="info")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

