% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbTranscriptClient-methods.R
\docType{methods}
\name{cbTranscriptClient,CellBaseR-method}
\alias{cbTranscriptClient}
\alias{cbTranscriptClient,CellBaseR-method}
\title{A method to query transcript data from Cellbase web services}
\usage{
\S4method{cbTranscriptClient}{CellBaseR}(object, ids, resource,
  filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried, for example,
ensemble transccript ID, like ENST00000380152}

\item{resource}{a character vector to specify the resource to be queried can
be any of "info", "function_prediction", "gene", "sequence", "variation"
, or "protein"}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query transcript data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbTranscriptClient(object=cb, ids="ENST00000373644", resource="info")
}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

