% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbVariantClient-methods.R
\docType{methods}
\name{cbVariantClient,CellBaseR-method}
\alias{cbVariantClient}
\alias{cbVariantClient,CellBaseR-method}
\title{A method to query variant annotation data from Cellbase web services from Cellbase web services.}
\usage{
\S4method{cbVariantClient}{CellBaseR}(object, ids, resource, filters = NULL,
  ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried, must be in the following format 'chr:start:ref:alt', for 
example, '1:128546:A:T'}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters for the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query variant annotation data from Cellbase web services from Cellbase web services.
}
\details{
Please, for details on possible values for the parameters  and  additional filters of this function refer to
https://github.com/opencb/cellbase/wiki and the RESTful API documentation 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbVariantClient(object=cb, ids="19:45411941:T:C", resource="annotation")
}

