% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbXref,CellbaseQuery-method}
\alias{cbXref}
\alias{cbXref,CellbaseQuery-method}
\title{A method to query cross reference data from Cellbase web services}
\usage{
\S4method{cbXref}{CellbaseQuery}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the ids to be queried can 
be any of "xref", "gene", "starts"with", or "contains"}

\item{filters}{a object of class CellbaseParam specifying additional filters
for the CellbaseQuery}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query cross reference data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellbaseQuery()
   res <- cbXref(object=cb, ids="ENST00000373644", resource="xref")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

