% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cbXrefClient,CellBaseR-method}
\alias{cbXrefClient}
\alias{cbXrefClient,CellBaseR-method}
\title{A method to query cross reference data from Cellbase web services}
\usage{
\S4method{cbXrefClient}{CellBaseR}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a charcter vector of the ids to be queried, any crossrefereable 
ID, gene names, transcript ids, uniprot ids,etc.}

\item{resource}{a charcter vector to specify the ids to be queried can 
be any of "xref", "gene", "starts"with", or "contains"}

\item{filters}{a object of class CellBaseParam specifying additional filters
for the CellBaseR}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
A method to query cross reference data from Cellbase web services
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   res <- cbXrefClient(object=cb, ids="ENST00000373644", resource="xref")
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

