% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cbXrefClient}
\alias{cbXrefClient}
\title{A method to query cross reference data from Cellbase web services. Please, for details on possible values for the 
parameters  and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
\usage{
cbXrefClient(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried, any crossrefereable ID, gene names, transcript ids, 
uniprot ids,etc.}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters for the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query cross reference data from Cellbase web services. Please, for details on possible values for the 
parameters  and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   library(cellbaseR)
   cb <- CellBaseR()
   res <- cbXrefClient(object=cb, ids="ENST00000373644", resource="xref")
}

