% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{cellbaseData,CellbaseResult-method}
\alias{cellbaseData,CellbaseResult-method}
\title{A method to query Clinical data from Cellbase web services}
\usage{
\S4method{cellbaseData}{CellbaseResult}(object)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{a object of class CellbaseParam specifying additional filterss for the CellbaseQuery}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
A method to query Clinical data from Cellbase web services
}

