% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{CellBaseParam}
\alias{CellBaseParam}
\title{A constructor function for CellBaseParam}
\usage{
CellBaseParam(genome = character(), gene = character(),
  region = character(), rs = character(), so = character(),
  phenotype = character(), include = character(), exclude = character(),
  limit = character())
}
\arguments{
\item{genome}{A character denoting the genome build to query,eg, GRCh37
(default),or GRCh38}

\item{gene}{A charcter vector denoting the gene/s to be queried}

\item{region}{A charcter vector denoting the region/s to be queried must be 
in the form 1:100000-1500000 not chr1:100000-1500000}

\item{rs}{A charcter vector denoting the rs ids to be queried}

\item{so}{A charcter vector denoting sequence ontology to be queried}

\item{phenotype}{A charcter vector denoting the phenotype to be queried}

\item{include}{A charcter vector denoting the fields to be returned}

\item{exclude}{A charcter vector denoting the fields to be excluded}

\item{limit}{A number limiting the number of results to be returned}
}
\description{
A constructor function for CellBaseParam
}
\details{
use the CellBaseParam object to control what results are returned from the 
CellBaseR methods
}
\examples{
library(cellbaseR)
cbParam <- CellBaseParam(genome="GRCh38",gene=c("TP73","TET1"))
print(cbParam)
}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

