% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{class}
\name{CellbaseQuery-class}
\alias{CellbaseQuery}
\alias{CellbaseQuery-class}
\title{This class defines the CellbaseQuery object which holds the default
configuration for connecting to the cellbase web services}
\description{
This class defines the CellbaseQuery object which holds the default
configuration for connecting to the cellbase web services
}
\details{
This class defines the CellbaseQuery object which holds the default
configuration required by CellbaseQuery methods to connect to the
cellbase web services. By defult it is configured to query human
data based on the GRCh37 genome assembly
}

