% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{CellbaseQuery}
\alias{CellbaseQuery}
\title{This is a constructor function for CellbaseQuery object which holds the default
configuration for connecting to the cellbase web services}
\usage{
CellbaseQuery(species = character())
}
\arguments{
\item{species}{A charcter should be one of the species supported by cellbase run cbSpecies
to see avaiable species and their corresponding data}
}
\value{
An object of class CellbaseQuery
}
\details{
This class defines the CellbaseQuery object which holds the default
configuration required by CellbaseQuery methods to connect to the
cellbase web services. By defult it is configured to query human
data based on the GRCh37 genome assembly
}
\section{Slots}{

\describe{
\item{\code{host}}{A character the default host url for cellllbase webservices}
}}
\examples{
library(cellbaseR)
 cb <- CellbaseQuery()
 print(cb)
}

