% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{class}
\name{CellBaseR-class}
\alias{CellBaseR-class}
\title{This class defines the CellnBaseR object}
\details{
This S4 class holds the default configuration required by
CellBaseR methods to connect to the cellbase web services.
By defult it is configured to query human
data based on the GRCh37 genome assembly
}
\section{Slots}{

\describe{
\item{\code{host}}{a charcter specifing the host url}

\item{\code{version}}{a charcter specifing the API version}

\item{\code{species}}{a character specifing the species to be queried}

\item{\code{batch_size}}{numeric a parameter for}

\item{\code{num_threads}}{the number of threads}
}}

