% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{CellBaseR}
\alias{CellBaseR}
\title{A constructor function for CellBaseR object}
\usage{
CellBaseR(species = character())
}
\arguments{
\item{species}{A charcter should be     a species supported by cellbase
run cbSpeciesClient to see avaiable species and their corresponding data}

\item{host}{A character the default host url for cellllbase webservices}

\item{version}{A character the cellbae API version}

\item{species}{A charcter should be     a species supported by cellbase}

\item{batch_size}{numeric a parameter for}
}
\value{
An object of class CellBaseR
}
\details{
This class defines the CellBaseR object. It holds the default
configuration required by CellBaseR methods to connect to the
cellbase web services. By defult it is configured to query human
data based on the GRCh37 genome assembly
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   print(cb)
}

