% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{class}
\name{CellBaseResponse-class}
\alias{CellBaseResponse}
\alias{CellBaseResponse-class}
\title{The CellBaseResponse class defintion}
\details{
This class stores the CellBaseR query methods, an object of
class CellBaseResponse is automatically generated when you call any of
CellbaseR methods
}
\section{Slots}{

\describe{
\item{\code{cbData}}{an R dataframe}
}}

