% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{createGeneModel2}
\alias{createGeneModel2}
\title{A function to plot gene modells from gene data returned by cbGeneClient}
\usage{
createGeneModel2(object, region = NULL)
}
\arguments{
\item{object}{an object of class CellBaseResponse}

\item{region}{a character}

\item{object}{an object of class CellbaseResponse}
}
\value{
A Gviz plot

A geneModel
}
\description{
A function to plot gene modells from gene data returned by cbGeneClient
}
\details{
This function takes cbResponse object and returns a geneRegionTrack
model to be plotted by Gviz
}

