% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getCaddScores}
\alias{getCaddScores}
\title{getCaddScores}
\usage{
getCaddScores(object, id, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{id}{a charcter vector of genomic variants, eg 19:45411941:T:C}

\item{param}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch Cadd scores for specific variant/s
}
\examples{
cb <- CellBaseR()
res <- getCaddScores(cb, "19:45411941:T:C")
}
