% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{getCellbase,CellbaseQuery-method}
\alias{getCellbase,CellbaseQuery-method}
\title{This method allows the user to query the cellbase web services without
any predefined categories, subcategries, or resources}
\usage{
\S4method{getCellbase}{CellbaseQuery}(object, categ, subcateg, ids, resource,
  filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellbaseQuery}

\item{categ}{charcter to specify the category to be queried
this could be "feature", "genomic", "regulatory", or "network"}

\item{subcateg}{charcter to specify the subcategory to be queried}

\item{ids}{a charcter vector of the ids to be queried}

\item{resource}{a charcter vector to specify the ids to be queried}

\item{filters}{an object of class CellbaseParam specifying additional filterss for the CellbaseQuery}

\item{...}{any extra arguments}

\item{file}{a path to a bgzipped and tabix indexed vcf file,}
}
\value{
an object of class CellbaseResult which holds a dataframe
}
\description{
This method allows the user to query the cellbase web services without
any predefined categories, subcategries, or resources
}

