% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{getCellbase,CellBaseR-method}
\alias{This}
\alias{allows}
\alias{any}
\alias{categories,}
\alias{cellbase}
\alias{getCellbase}
\alias{getCellbase,CellBaseR-method}
\alias{method}
\alias{or}
\alias{predefined}
\alias{query}
\alias{resources}
\alias{services}
\alias{subcategries,}
\alias{the}
\alias{to}
\alias{user}
\alias{web}
\alias{without}
\title{The generic method for getCellbase}
\usage{
\S4method{getCellbase}{CellBaseR}(object, file = NULL, category, subcategory,
  ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{file}{a path to a bgzipped and tabix indexed vcf file,}

\item{category}{character to specify the category to be queried
this could be "feature", "genomic", "regulatory", or "network"}

\item{subcategory}{character to specify the subcategory to be queried}

\item{ids}{a character vector of the ids to be queried}

\item{resource}{a character to specify the resource to be queried}

\item{filters}{an object of class CellBaseParam specifying additional 
filterss for the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
The generic method for getCellbase
}
\examples{
   library(cellbaseR)
   cb <- CellBaseR
   res <- getCellbase(object=cb,categ="feature",subcateg="gene",ids="TET1",
   resource="info")

}
\seealso{
for more information about the cellbase webservices see 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

