% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getCellbase}
\alias{getCellbase}
\title{The generic method for getCellbase. This method allows the user to query the cellbase web services without any 
predefined categories, subcategries, or resources. Please, for details on possible values for the parameters and 
additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful API documentation 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
\usage{
getCellbase(object, categ, subcateg, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried}

\item{resource}{a character to specify the resource to be queried}

\item{filters}{an object of class CellBaseParam specifying additional filters for the CellBaseR}

\item{...}{any extra arguments}

\item{category}{character to specify the category to be queried.}

\item{subcategory}{character to specify the subcategory to be queried}
}
\value{
an object of class CellBaseResponse which holds a dataframe
}
\description{
The generic method for getCellbase. This method allows the user to query the cellbase web services without any 
predefined categories, subcategries, or resources. Please, for details on possible values for the parameters and 
additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful API documentation 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   library(cellbaseR)
   cb <- CellBaseR
   res <- getCellbase(object=cb,categ="feature",subcateg="gene",ids="TET1",
   resource="info")
}
\seealso{
for more information about the cellbase webservices see
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}

