% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{getGeneInfo}
\usage{
getGeneInfo(object, id, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{id}{a charcter vector of HUGO symbol (gene names)}

\item{param}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch gene annotations specific gene/s
}
\examples{
cb <- CellBaseR()
res <- getGeneInfo(cb, "TET1")
}
