% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta-methods.R
\docType{methods}
\name{getMeta,CellBaseR-method}
\alias{getMeta,CellBaseR-method}
\alias{getMeta}
\title{getMeta}
\usage{
\S4method{getMeta}{CellBaseR}(object, resource)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{resource}{the resource you want to query it metadata}
}
\value{
a dataframe with the
 results of the query
}
\description{
A method for getting the available metadata from the cellbase web services
}
\details{
This method is for getting information about the avaialable species
and available annotation, assembly for each species from the cellbase web 
services.
}
\examples{
   cb <- CellBaseR()
   res <- getMeta(object=cb, resource="species")
}
\seealso{
\url{https://github.com/opencb/cellbase/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}
