% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getProteinInfo}
\alias{getProteinInfo}
\title{getProteinInfo}
\usage{
getProteinInfo(object, id, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{id}{a charcter vector of Uniprot Ids}

\item{param}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch annotations for specific protein/s
}
\examples{
cb <- CellBaseR()
res <- getProteinInfo(cb, "O15350")
}
