% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getRegulatoryByRegion}
\alias{getRegulatoryByRegion}
\title{getRegulatoryByRegion}
\usage{
getRegulatoryByRegion(object, id, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{id}{a charcter vector of genomic regions, eg 17:1000000-1100000}

\item{param}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch regulatory data for specific region/s
}
\examples{
cb <- CellBaseR()
res <- getRegulatoryByRegion(cb, "17:1000000-1189811")
}
