% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getSnpByGene}
\alias{getSnpByGene}
\title{getSnpByGene}
\usage{
getSnpByGene(object, id, param = NULL)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{id}{a charcter vector of HUGO symbol (gene names)}

\item{param}{an object of class CellBaseParam}
}
\value{
a dataframe of the query result
}
\description{
A convienice method to fetch known variants (snps) for specific gene/s
}
\examples{
cb <- CellBaseR()
param <- CellBaseParam(limit = 10)
res <- getSnpByGene(cb, "TET1", param = param)
}
