% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{host<-}
\alias{host<-}
\title{A method to allow for updating the host of the CellbaseQuery config}
\usage{
host(object) <- value
}
\arguments{
\item{object}{An object of class CellbaseQuery}

\item{value}{a character specifying the desired host}
}
\description{
A method to allow for updating the host of the CellbaseQuery config
}

