% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{plotGenes}
\alias{plotGenes}
\title{A function to plot gene modells from gene data returned by cbGeneClient}
\usage{
plotGenes(object)
}
\arguments{
\item{object}{an object of class CellBaseResponse}
}
\value{
A Gviz plot
}
\description{
A function to plot gene modells from gene data returned by cbGeneClient
}

