% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,CellBaseR-method}
\alias{show,CellBaseR-method}
\title{The show method for CellBaseR class}
\usage{
\S4method{show}{CellBaseR}(object)
}
\arguments{
\item{object}{an object of class CellBaseR}
}
\description{
The show method for CellBaseR class
}

