% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,CellBaseResponse-method}
\alias{show,CellBaseResponse-method}
\title{The show method for CellBaseResponse class}
\usage{
\S4method{show}{CellBaseResponse}(object)
}
\arguments{
\item{object}{an object of class CellBaseResponse}
}
\description{
The show method for CellBaseResponse class
}

