% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{species<-,CellbaseQuery-method}
\alias{species<-,CellbaseQuery-method}
\title{A method to allow for updating the species of the CellbaseQuery config}
\usage{
\S4method{species}{CellbaseQuery}(object) <- value
}
\arguments{
\item{object}{An object of class CellbaseQuery}

\item{value}{a character specifying the desired species}
}
\description{
A method to allow for updating the species of the CellbaseQuery config
}

