% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{species<-}
\alias{species<-}
\title{A method to allow for updating the species of the CellbaseQuery config}
\usage{
species(object) <- value
}
\description{
A method to allow for updating the species of the CellbaseQuery config
}

