% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbData-methods.R
\docType{methods}
\name{[,CellBaseResponse,ANY,ANY-method}
\alias{[,CellBaseResponse,ANY,ANY-method}
\title{A substter method for CellBaseResponse class objects
A substter method for CellBaseResponse class objects}
\usage{
\S4method{[}{CellBaseResponse,ANY,ANY}(x, i, j, drop = "missing")
}
\arguments{
\item{i, j, }{indices specifying elements to extract or replace. Indices are numeric
or character vectors or empty (missing)}

\item{object}{from which to extract element(s).}
}

