% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\docType{methods}
\name{[,CellbaseResult,ANY,ANY-method}
\alias{[,CellbaseResult,ANY,ANY-method}
\title{A method to allow for subsetting the cellbaseData}
\usage{
\S4method{[}{CellbaseResult,ANY,ANY}(x, i, j, drop = "missing")
}
\description{
A method to allow for subsetting the cellbaseData
}

