% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbAnnotateVcf-methods.R
\docType{methods}
\name{cbAnnotateVcf,CellBaseR-method}
\alias{cbAnnotateVcf}
\alias{cbAnnotateVcf,CellBaseR-method}
\title{This method is a convience method to annotate a vcf files. This methods is ideal for annotating small to medium sized 
vcf files.}
\usage{
\S4method{cbAnnotateVcf}{CellBaseR}(object, file, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{file}{Path to a bgzipped and tabix indexed vcf file}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
This method is a convience method to annotate a vcf files. This methods is ideal for annotating small to medium sized 
vcf files.
}

