% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbClinicalClient-methods.R
\docType{methods}
\name{cbClinicalClient,CellBaseR-method}
\alias{cbClinicalClient}
\alias{cbClinicalClient,CellBaseR-method}
\title{A method to query Clinical data from Cellbase web services.}
\usage{
\S4method{cbClinicalClient}{CellBaseR}(object, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{filters}{a object of class CellBaseParam specifying the parameters limiting the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query Clinical data from Cellbase web services.
}
\details{
Please, for details on possible values for the parameters 
and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful API 
documentation  http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   library(cellbaseR)
   cb <- CellBaseR()
   cbParam <- CellBaseParam(gene=c("TP73","TET1"))
   res <- cbClinicalClient(object=cb,filters=cbParam)
}

