% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbGeneClient-methods.R
\docType{methods}
\name{cbGeneClient,CellBaseR-method}
\alias{cbGeneClient}
\alias{cbGeneClient,CellBaseR-method}
\title{A method to query gene data from Cellbase web services.}
\usage{
\S4method{cbGeneClient}{CellBaseR}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of gene ids to be queried}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{an object of class CellBaseParam specifying additional filters
for the CellBaseR}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the 
results of the query
}
\description{
A method to query gene data from Cellbase web services.
}
\details{
Please, for details on possible values for the parameters 
and  additional filters of this function refer to
 https://github.com/opencb/cellbase/wiki and the RESTful API 
documentation  http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbGeneClient(object=cb, ids=c("TP73","TET1"), resource="info")
}

