% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRegionClient-methods.R
\docType{methods}
\name{cbRegionClient,CellBaseR-method}
\alias{cbRegionClient}
\alias{cbRegionClient,CellBaseR-method}
\title{A method to query features within a genomic region from Cellbase web services.}
\usage{
\S4method{cbRegionClient}{CellBaseR}(object, ids, resource, filters = NULL,
  ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the regions to be queried, for example, "1:1000000-1200000' should always be in the
form 'chr:start-end'}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters for the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query features within a genomic region from Cellbase web services.
}
\details{
Please, for details on possible values
for the parameters  and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the 
RESTful API documentation  http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbRegionClient(object=cb, ids="17:1000000-1200000", resource="gene")
}

