% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbTfbsClient-methods.R
\docType{methods}
\name{cbTfbsClient,CellBaseR-method}
\alias{cbTfbsClient}
\alias{cbTfbsClient,CellBaseR-method}
\title{A method to query transcription factors binding sites data from Cellbase web services.}
\usage{
\S4method{cbTfbsClient}{CellBaseR}(object, ids, resource, filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the ids to be queried, must be a valid transcription factor name, for example, 
Pdr1, and Oaf1}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters for the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query transcription factors binding sites data from Cellbase web services.
}
\details{
Please for details on possible values for the parameters  and  additional filters of this function refer to
https://github.com/opencb/cellbase/wiki and the RESTful API documentation 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbTfbsClient(object=cb, ids="PAX1", resource="gene")
}

