% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbTranscriptClient-methods.R
\docType{methods}
\name{cbTranscriptClient,CellBaseR-method}
\alias{cbTranscriptClient}
\alias{cbTranscriptClient,CellBaseR-method}
\title{A method to query transcript data from Cellbase web services.}
\usage{
\S4method{cbTranscriptClient}{CellBaseR}(object, ids, resource,
  filters = NULL, ...)
}
\arguments{
\item{object}{an object of class CellBaseR}

\item{ids}{a character vector of the transcript ids to be queried, for example, ensemble transccript ID
like ENST00000380152}

\item{resource}{a character vector to specify the resource to be queried}

\item{filters}{a object of class CellBaseParam specifying additional filters for the query}

\item{...}{any extra arguments}
}
\value{
an object of class CellBaseResponse which holds a dataframe with the results of the query
}
\description{
A method to query transcript data from Cellbase web services.
}
\details{
Please, for details on possible values for the 
parameters  and  additional filters of this function refer to https://github.com/opencb/cellbase/wiki and the RESTful 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/
}
\examples{
   cb <- CellBaseR()
   res <- cbTranscriptClient(object=cb, ids="ENST00000373644", resource="info")
}

