% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellBaseParam-class}
\alias{CellBaseParam-class}
\title{This Class defines a CellBaseParam object}
\description{
This class  defines a CellBaseParam object to hold filtering parameters
}
\details{
This class stores filtering parameters to be used by CellBaseR query
 methods. Not all the slots will be
used by all query methods, please have a look at 
http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/ and the 
Reference Manual for more information.
}
\section{Slots}{

\describe{
\item{\code{genome}}{A character the genome build to query, e.g.GRCh37(default)}

\item{\code{gene}}{A character vector denoting the gene/s to be queried}

\item{\code{region}}{A character vector denoting the region/s to be queried must be 
in the form 1:100000-1500000}

\item{\code{rs}}{A character vector denoting the rs ids to be queried}

\item{\code{so}}{A character vector denoting sequence ontology to be queried}

\item{\code{phenotype}}{A character vector denoting the phenotype to be queried}

\item{\code{include}}{A character vector denoting the fields to be returned}

\item{\code{exclude}}{A character vector denoting the fields to be excluded}

\item{\code{limit}}{A number limiting the number of results to be returned}
}}

