% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{CellbaseQuery}
\alias{CellbaseQuery}
\title{This is a constructor function for CellbaseQuery object}
\usage{
CellbaseQuery(species = character())
}
\arguments{
\item{species}{A charcter should be     a species supported by cellbase
run cbSpecies to see avaiable species and their corresponding data}
}
\value{
An object of class CellbaseQuery
}
\details{
This class defines the CellbaseQuery object. It holds the default
configuration required by CellbaseQuery methods to connect to the
cellbase web services. By defult it is configured to query human
data based on the GRCh37 genome assembly
}
\section{Slots}{

\describe{
\item{\code{host}}{A character the default host url for cellllbase webservices}

\item{\code{version}}{A character the cellbae API version}

\item{\code{species}}{A charcter should be     a species supported by cellbase}

\item{\code{batch_size}}{numeric a parameter for}
}}
\examples{
library(cellbaseR)
   cb <- CellbaseQuery()
   print(cb)
}

