% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBaseR-methods.R
\name{CellBaseR}
\alias{CellBaseR}
\title{A constructor function for CellBaseR object}
\usage{
CellBaseR(host = NULL, version = NULL, species = NULL,
  batch_size = NULL, num_threads = NULL)
}
\arguments{
\item{host}{A character the default host url for cellbase webservices,
e.g. "http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/rest/"}

\item{version}{A character the cellbae API version, e.g. "V4"}

\item{species}{a character specifying the species to be queried, e.g. "hsapiens"}

\item{batch_size}{intger if multiple queries are raised by a single method call, e.g. getting annotation info for several genes,
queries will be sent to the server in batches. This slot indicates the size of each batch, e.g. 200}

\item{num_threads}{integer number of asynchronus batches to be sent to the server}
}
\value{
An object of class CellBaseR
}
\description{
This is a constructor function for CellBaseR object
}
\details{
This class defines the CellBaseR object. It holds the default
configuration required by CellBaseR methods to connect to the
cellbase web services. By defult it is configured to query human
data based on the GRCh37 genome assembly. Please, visit
https://github.com/opencb/cellbase/wiki and
bioinfodev.hpc.cam.ac.uk/cellbase/webservices/ for more details on
following parameters.
run cbSpeciesClient to see avaiable species and their corresponding data
}
\examples{
library(cellbaseR)
   cb <- CellBaseR()
   print(cb)
}

