% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellBaseResponse-class}
\alias{CellBaseResponse}
\alias{CellBaseResponse-class}
\title{The CellBaseResponse class defintion
 
This class holds the response data from CellBaseR Methods}
\description{
The CellBaseResponse class defintion
 
This class holds the response data from CellBaseR Methods
}
\details{
This class stores a response of CellBaseR query methods. An object 
of class CellBaseResponse is automatically generated when you call any of 
CellbaseR methods.
}
\section{Slots}{

\describe{
\item{\code{cbData}}{an R dataframe which contains the result field within the 
response object returned by CellBase web services.}
}}

