% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createGeneModel}
\alias{createGeneModel}
\title{createGeneModel}
\usage{
createGeneModel(object, region = NULL)
}
\arguments{
\item{object}{an object of class CellbaseResponse}

\item{region}{a character}
}
\value{
A geneModel
}
\description{
A convience functon to construct a genemodel
}
\details{
This function create a gene model data frame, which can be then 
turned into a GeneRegionTrack for visualiaztion
 by \code{\link[Gviz]{GeneRegionTrack}}
}
\examples{
cb <- CellBaseR()
test <- createGeneModel(object = cb, region = "17:1500000-1550000")
}
\seealso{
\url{https://github.com/opencb/cellbase/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/}
}
