% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbase_classes.R
\name{getCellbase}
\alias{getCellbase}
\title{The generic method for getCellbase}
\usage{
getCellbase(object, file = NULL, categ, subcateg, ids, resource,
  filters = NULL, ...)
}
\description{
This method allows the user to query the cellbase web services without
any predefined categories, subcategries, or resources
}

